I'd like to welcome you to...
I'd like to welcome you to the icCoralServer, a lightweight server which uses the infinite campus API to sign in and verify that it is you and then sends the information back to anyone who wields the access token. Will add styling to make it a better site since the page for verification looks very bad.

The end-plan is for this to be able to connect to a Discord bot in the server for you to verify that you are one of the students and therefore gain access to the comms and information.